web-team-dashboard/backend/requirements.txt
Django==4.2.24
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
gunicorn==23.0.0
django-filter==24.3
Pillow==11.0.0